re-name handlers and tests for consistentcy
