+++ handlers/timer_start.js
