+++ test/timer_start.js
